+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackendImpl.kt
@@ -19,7 +19,9 @@ import cash.z.ecc.android.sdk.model.AccountPurpose
 import cash.z.ecc.android.sdk.model.AccountUsk
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FirstClassByteArray
+import cash.z.ecc.android.sdk.model.Pczt
 import cash.z.ecc.android.sdk.model.Proposal
+import cash.z.ecc.android.sdk.model.UnifiedFullViewingKey
 import cash.z.ecc.android.sdk.model.UnifiedSpendingKey
 import cash.z.ecc.android.sdk.model.Zatoshi
 import cash.z.ecc.android.sdk.model.ZcashNetwork
@@ -84,6 +86,10 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
         )
     }
 
+    override suspend fun getAccountForUfvk(ufvk: UnifiedFullViewingKey): Account? {
+        return backend.getAccountForUfvk(ufvk = ufvk.encoding)?.let { Account.new(it) }
+    }
+
     override suspend fun proposeTransferFromUri(
         account: Account,
         uri: String
@@ -136,6 +142,30 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
             proposal.toUnsafe(),
             usk.copyBytes()
         ).map { FirstClassByteArray(it) }
+
+    override suspend fun createPcztFromProposal(
+        account: Account,
+        proposal: Proposal
+    ): Pczt =
+        Pczt(
+            backend.createPcztFromProposal(
+                account.accountUuid.value,
+                proposal.toUnsafe()
+            )
+        )
+
+    override suspend fun addProofsToPczt(pczt: Pczt): Pczt = Pczt(backend.addProofsToPczt(pczt.toByteArray()))
+
+    override suspend fun extractAndStoreTxFromPczt(
+        pcztWithProofs: Pczt,
+        pcztWithSignatures: Pczt
+    ): FirstClassByteArray =
+        FirstClassByteArray(
+            backend.extractAndStoreTxFromPczt(
+                pcztWithProofs.toByteArray(),
+                pcztWithSignatures.toByteArray()
+            )
+        )
 
     override suspend fun getCurrentAddress(account: Account): String {
         return runCatching {